+++ pymadagascar/seismic/__init__.py
@@ -2,11 +2,13 @@
 
 from .ai2refl import AI2ReflError, ai2refl, ai2refl_rsf
 from .agc import AGCError, agc_rsf
+from .angle import AngleTransformError, cos2ang, cos2ang_rsf, isin2ang, isin2ang_rsf
 from .avo import AVOError, avo_intercept_gradient, avo_rsf
 from .fk import FKError, fk_filter, fk_spectrum, make_fk_mask
 from .fold import FoldError, fold_rsf, fold_table
 from .halfint import HalfIntError, halfint, halfint_rsf
 from .gain import GainError, gain_rsf
+from .map2coh import Map2CohError, map2coh, map2coh_rsf
 from .moveout import MoveoutError, moveout_rsf, moveout_spikes
 from .mute import MuteError, mute_rsf, mutter, mutter_rsf
 from .nmo import NMOError, inverse_nmo, nmo_correct
@@ -25,11 +27,13 @@ from .stack import StackError, stack_along_axis, stack_rsf, stacks_rsf
 __all__ = [
     "AGCError",
     "AI2ReflError",
+    "AngleTransformError",
     "AVOError",
     "FKError",
     "FoldError",
     "GainError",
     "HalfIntError",
+    "Map2CohError",
     "MuteError",
     "MoveoutError",
     "NMOError",
@@ -41,6 +45,8 @@ __all__ = [
     "agc_rsf",
     "avo_intercept_gradient",
     "avo_rsf",
+    "cos2ang",
+    "cos2ang_rsf",
     "fk_filter",
     "fk_spectrum",
     "fold_rsf",
@@ -48,11 +54,15 @@ __all__ = [
     "gain_rsf",
     "halfint",
     "halfint_rsf",
+    "isin2ang",
+    "isin2ang_rsf",
     "inverse_nmo",
     "inverse_linear_radon",
     "inverse_parabolic_radon",
     "linear_radon",
     "make_fk_mask",
+    "map2coh",
+    "map2coh_rsf",
     "mute_rsf",
     "mutter",
     "mutter_rsf",